// packages/api/src/types/todo/adx/deleteAccount.ts
/**
* GENERATED CODE - DO NOT MODIFY
* Created Wed Sep 21 2022
*/
import { Headers } from '@adxp/xrpc'

export interface QueryParams {}

export interface CallOptions {
  headers?: Headers;
  encoding: '';
}

export interface InputSchema {
  [k: string]: unknown;
}

export interface OutputSchema {
  [k: string]: unknown;
}

export interface Response {
  success: boolean;
  error: boolean;
  headers: Headers;
  data: OutputSchema;
}